--- packages/lib-classifier/src/components/Classifier/components/Layout/components/DefaultLayout/index.js
@@ -1,1 +1,0 @@
-export { default } from './DefaultLayout'
