+++ packages/lib-classifier/src/components/Classifier/components/Layout/components/NoMaxWidth/index.js
@@ -1,0 +1,1 @@
+export { default } from './NoMaxWidth'
