+++ packages/lib-classifier/src/components/Classifier/components/Layout/components/index.js
@@ -1,2 +1,2 @@
-export { default as DefaultLayout } from './DefaultLayout'
 export { default as MaxWidth } from './MaxWidth'
+export { default as NoMaxWidth } from './NoMaxWidth'
