+++ packages/lib-classifier/src/components/Classifier/components/Layout/helpers/getLayout/getLayout.spec.js
@@ -1,17 +1,17 @@
 import getLayout from './getLayout'
 
-import { DefaultLayout, MaxWidth } from '../../components'
+import { NoMaxWidth, MaxWidth } from '../../components'
 
 describe('Helpers > getLayout', function () {
-  it('should return the `MaxWidth` component if passed `default`', function () {
+  it('should return the `MaxWidth` component if passed `maxWidth`', function () {
     expect(getLayout('default')).to.equal(MaxWidth)
   })
 
-  it('should return the old `DefaultLayout` component if passed `noMaxWidth`', function () {
-    expect(getLayout('noMaxWidth')).to.equal(DefaultLayout)
+  it('should return the `NoMaxWidth` component if passed `noMaxWidth`', function () {
+    expect(getLayout('noMaxWidth')).to.equal(NoMaxWidth)
   })
 
-  it('should return the default layout if it can\'t match a layout', function () {
+  it('should return the `MaxWidth` layout if it can\'t match a layout', function () {
     expect(getLayout('foobar')).to.equal(MaxWidth)
   })
 })
